Vincent's block code
We believe after we tested all of our codes and through collaboration, Ransom's post was our final version of the code. It is our final version of the code since it has the least amount of steps and the code we tested worked flawlessly. My code contained a loop and the code did not allow the robot to follow the correct path and went all over the place.